Utilities for testing runtime functions.
